Move Lua reference helpers for #3421
